+++ lib/widgets/components/stats/leaderboard_list_entry.dart
@@ -3,7 +3,7 @@ import 'package:coffeecard/base/style/colors.dart';
 import 'package:coffeecard/base/style/text_styles.dart';
 import 'package:coffeecard/widgets/components/helpers/shimmer_builder.dart';
 import 'package:coffeecard/widgets/components/list_entry.dart';
-import 'package:coffeecard/widgets/components/user_icon.dart';
+import 'package:coffeecard/widgets/components/user/user_icon.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
