+++ lib/widgets/components/user/user_card.dart
@@ -2,7 +2,7 @@ import 'package:coffeecard/base/style/colors.dart';
 import 'package:coffeecard/base/style/text_styles.dart';
 import 'package:coffeecard/widgets/components/helpers/shimmer_builder.dart';
 import 'package:coffeecard/widgets/components/helpers/tappable.dart';
-import 'package:coffeecard/widgets/components/user_icon.dart';
+import 'package:coffeecard/widgets/components/user/user_icon.dart';
 import 'package:coffeecard/widgets/pages/settings/your_profile_page.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
